+++ gui/src/tabs/web/__init__.py
@@ -1,4 +1,5 @@
 from .drive_sync_tab import DriveSyncTab as DriveSyncTab
+from .entity_recon_tab import EntityReconTab as EntityReconTab
 from .image_crawler_tab import ImageCrawlTab as ImageCrawlTab
 from .reverse_search_tab import ReverseImageSearchTab as ReverseImageSearchTab
 from .web_requests_tab import WebRequestsTab as WebRequestsTab